Fixing issue #4
Problem: Fix failed deletes of heat stacks when resources fail

Analysis: When deleting a heat stack, resources can fail to delete and
the stack is permanently broken. The fix is to delete f5 resources in a
smarter way, that is, check if that resource exists on the BigIP device,
and if it does, delete it. Otherwise, no action needs to be taken.

Unit/Robot tests: All tests pass and we have 100% coverage

+++ f5_heat/resources/f5_ltm_pool.py
@@ -141,14 +141,19 @@ class F5LTMPool(resource.Resource, F5BigIPMixin):
         :raises: ResourceFailure
         '''
 
-        try:
-            loaded_pool = self.bigip.ltm.pools.pool.load(
+        if self.bigip.ltm.pools.pool.exists(
                 name=self.properties[self.NAME],
                 partition=self.partition_name
-            )
-            loaded_pool.delete()
-        except Exception as ex:
-            raise exception.ResourceFailure(ex, None, action='DELETE')
+        ):
+            try:
+                loaded_pool = self.bigip.ltm.pools.pool.load(
+                    name=self.properties[self.NAME],
+                    partition=self.partition_name
+                )
+                loaded_pool.delete()
+            except Exception as ex:
+                raise exception.ResourceFailure(ex, None, action='DELETE')
+        return True
 
 
 def resource_mapping():
